Refactored transforming to separate trait
